requirements.txt
pydantic[email]
fastapi==0.88.0
uvicorn==0.20.0
sqlalchemy[asyncio]==1.4.45
python-jose==3.3.0
python-multipart==0.0.5
alembic
asyncpg
passlib
